tests: impl unit tests for animals controller

+++ VetClinicShelterApi/VetClinicShelterApi/Dtos/Response/AnimalResponseDto.cs
@@ -3,4 +3,4 @@ using VetClinicShelterApi.Models;
 
 namespace VetClinicShelterApi.Dtos.Response;
 
-public record AnimalResponseDto(Guid Id, string Name, decimal Weight, AnimalCategory Category, string Color);
+public record AnimalResponseDto(Guid Id, string Name, double Weight, AnimalCategory Category, string Color);
